+++ catalog/controller/payment/nuvei.php
@@ -781,11 +781,11 @@ class Nuvei extends \Opencart\System\Engine\Controller
 		);
         
         // change urlDetails
-        if (!empty($this->plugin_settings[NUVEI_SETTINGS_PREFIX . 'auto_close_apm_popup'])) {
+//        if (!empty($this->plugin_settings[NUVEI_SETTINGS_PREFIX . 'auto_close_apm_popup'])) {
             $oo_params['urlDetails']['successUrl']  = $oo_params['urlDetails']['failureUrl']
                                                     = $oo_params['urlDetails']['pendingUrl']
                                                     = NUVEI_SDK_AUTOCLOSE_URL;
-        }
+//        }
         
         $oo_params = array_merge_recursive($oo_params, $rebilling_params);